+++ src/main/java/com/t2m/g2nee/shop/fileset/bookfile/domain/BookFile.java
@@ -1,0 +1,27 @@
+package com.t2m.g2nee.shop.fileset.bookfile.domain;
+
+import com.t2m.g2nee.shop.bookset.Book.domain.Book;
+import com.t2m.g2nee.shop.fileset.file.domain.File;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.experimental.SuperBuilder;
+
+import javax.persistence.*;
+
+
+@Entity
+@Table(name = "BookFiles")
+@Getter
+@Setter
+@SuperBuilder
+@NoArgsConstructor
+@AllArgsConstructor
+@DiscriminatorValue("BookFiles")
+public class BookFile extends File {
+
+    @ManyToOne
+    @JoinColumn(name = "bookId")
+    private Book book;
+}
