+++ src/main/java/com/t2m/g2nee/shop/fileset/file/domain/File.java
@@ -1,0 +1,26 @@
+package com.t2m.g2nee.shop.fileset.file.domain;
+
+
+import lombok.*;
+import lombok.experimental.SuperBuilder;
+
+import javax.persistence.*;
+
+@Inheritance(strategy = InheritanceType.JOINED)
+@DiscriminatorColumn
+@Entity
+@Table(name = "Files")
+@Getter
+@Setter
+@SuperBuilder
+@NoArgsConstructor
+@AllArgsConstructor
+public class File {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long fileId;
+
+    private String url;
+    private String extensionName;
+}
